requirements.txt
transformers
optimum
onnxruntime
langchain
langchain-community
langchain-huggingface
# langchain-core==0.3.15
pdfplumber
PyMuPDF
numpy
llama-cpp-python
onnx
sentence-transformers
chromadb